Rename save-hashes to save, add option --hash (default) and --branch.
