+++ lib/commands/save.js
@@ -40,7 +40,7 @@ module.exports = {
 		const options = this._parseArguments( args );
 
 		if ( !options.branch && !options.hash) {
-			throw new Error( 'Need to specify what kind of information you want to save. Call "mgit commit -h" in order to read more.' );
+			throw new Error( 'Need to specify what kind of information you want to save. Call "mgit save -h" in order to read more.' );
 		}
 	},
 
@@ -55,6 +55,7 @@ module.exports = {
 
 		let promise;
 
+		/* istanbul ignore else */
 		if ( options.branch ) {
 			promise =  execCommand.execute( getExecData( 'git status --branch --porcelain' ) )
 				.then( execResponse => gitStatusParser( execResponse.logs.info[ 0 ] ).branch );
@@ -71,6 +72,7 @@ module.exports = {
 				hash: options.hash
 			};
 
+			/* istanbul ignore else */
 			if ( options.branch ) {
 				log.info( `Branch: "${ dataToSave }".` );
 			} else if ( options.hash ) {
